add checkout user page feature

+++ app/src/main/java/com/taskapp/domain/Review.kt
@@ -5,7 +5,7 @@ import java.util.*
 data class Review(
     val reviewer_id: String? = null,
     val reviewee_id: String? = null,
-    val num_stars: Double? = null,
+    val num_stars: Float? = null,
     val comment: String? = null,
     val creation_date: Date? = null,
 )
